added service classes

+++ ERS-V2/back-end/ers-v2-backend/src/main/java/com/ersv2/repos/AddressRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ersv2.models.Address;
 
 @Repository
-public interface AddressRepo extends JpaRepository <Address, Integer>{
+public interface AddressRepo extends JpaRepository <Address, Long>{
 
 }
